+++ src/test_methods.js
@@ -52,6 +52,7 @@ async function runTests() {
   await testEndpoint('getDexPools', '/networks/ethereum/dexes/uniswap_v3/pools');
   await testEndpoint('getPoolDetails', '/networks/ethereum/pools/0x88e6a0c2ddd26feeb64f039a2c41296fcb3f5640');
   await testEndpoint('getTokenDetails', '/networks/ethereum/tokens/0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48');
+  await testEndpoint('getTokenMultiPrices', '/networks/ethereum/multi/prices?tokens=0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2&tokens=0xdac17f958d2ee523a2206206994597c13d831ec7');
   await testEndpoint('search', '/search?query=ethereum');
   await testEndpoint('getStats', '/stats');